+++ src/hooks/use-iiif/index.ts
@@ -7,12 +7,22 @@ import { getAccompanyingCanvasImage } from "src/hooks/use-iiif/getAccompanyingCa
 import { getCanvasByCriteria } from "src/hooks/use-iiif/getCanvasByCriteria";
 import { getLabel } from "src/hooks/use-iiif/getLabel";
 import { getPaintingResource } from "src/hooks/use-iiif/getPaintingResource";
+import {
+  getPlayerCaptions,
+  getPlayerChapters,
+  getPlayerResources,
+  getPlayerSources,
+} from "src/hooks/use-iiif/getPlayerResources";
 
 export {
   getAccompanyingCanvasImage,
   getCanvasByCriteria,
   getLabel,
   getPaintingResource,
+  getPlayerCaptions,
+  getPlayerChapters,
+  getPlayerResources,
+  getPlayerSources,
   getAnnotationResources,
   getContentSearchResources,
 };